optimization/optimizer.py
import dspy
import os
import json
from dspy.telemetry import *
from optimization.signatures import QueryRewriter, DocGrader, AnswerGenerator

class RAGAgent(dspy.Module):
    def __init__(self):
        super().__init__()
        self.rewriter = dspy.ChainOfThought(QueryRewriter)
        self.grader = dspy.Predict(DocGrader)
        self.generator = dspy.ChainOfThought(AnswerGenerator)

    def forward(self, question, context_list):
        # 1. Rewrite
        rewrite_res = self.rewriter(question=question)
        
        # 2. Grade (Assuming context is passed for simplicity in this baseline)
        context_str = "\n\n".join(context_list)
        grade_res = self.grader(question=question, context=context_str)
        
        # 3. Generate
        answer_res = self.generator(question=question, context=context_str)
        
        return dspy.Prediction(
            rewritten_query=rewrite_res.optimized_query,
            is_sufficient=grade_res.is_sufficient,
            answer=answer_res.answer
        )

# Example logic for optimization run
def setup_dspy():
    # Use Groq via DSPy's OpenAI-compatible adapter or a custom wrapper
    # For now, we'll scaffold with a mock/test config
    # In practice, user needs to set GROQ_API_KEY
    pass

if __name__ == "__main__":
    # Scaffolding for future optimization runs
    print("DSPy RAG Module initialized.")
