requirements.txt
gradio
fastapi
uvicorn
python-dotenv
groq
langgraph
sentence-transformers
qdrant-client
chromadb
gitpython
beautifulsoup4
markdown
pydantic
pytest
httpx
tiktoken
langgraph
langfuse
ragas
dspy-ai
ruff
slowapi
rank_bm25
python-multipart
